+++ Library/File/IFileDataRepository.cs
@@ -5,6 +5,6 @@ namespace Library.File
     public interface IFileDataRepository
     {
         BigInteger GetByteCount(string fileName);
-        byte[] GetPage(string fileName, ushort pageSize, BigInteger pageIndex);
+        byte[] GetPage(string fileName, uint pageSize, BigInteger pageIndex);
     }
 }
